+++ nodes/codec.js
@@ -34,7 +34,8 @@ const enc = new Map([
   [opcodes.GetEvent, encoder.GetEvent],
   [opcodes.SetPCControl, encoder.SetPCControl],
   [opcodes.SetInterlock, encoder.SetInterlock],
-  [opcodes.ActivateKeypads, encoder.ActivateKeypads]
+  [opcodes.ActivateKeypads, encoder.ActivateKeypads],
+  [opcodes.SetSuperPasswords, encoder.SetSuperPasswords]
 ])
 
 /**
@@ -59,6 +60,7 @@ const dec = new Map([
   [0x98, decoder.GetTimeProfile],
   [0x88, decoder.SetTimeProfile],
   [0x8a, decoder.ClearTimeProfiles],
+  [opcodes.SetSuperPasswords, decoder.SetSuperPasswords],
   [0x8e, decoder.RecordSpecialEvents],
   [0xa0, decoder.SetPCControl],
   [0xa2, decoder.SetInterlock],