+++ nodes/decoder.js
@@ -611,6 +611,22 @@ module.exports = {
       deviceId: uint32(bytes, 4),
       updated: bool(bytes, 8)
     }
+  },
+
+  /**
+    * Decodes the response to a set-super-passwords request (function code 0x8c).
+    *
+    * @param {buffer}   buffer     64 byte NodeJS buffer
+    * @param {function} translator (optional) function to internationalise the text in a
+    *                              decoded object
+    *
+    * @param {object}   Decoded set-super-passwords response object
+    */
+  SetSuperPasswords: function (bytes, translator) {
+    return {
+      deviceId: uint32(bytes, 4),
+      ok: bool(bytes, 8)
+    }
   }
 }